iox-#2278 Split logstream tests into multiple files
