+++ iceoryx_hoofs/test/moduletests/test_reporting_logstream.cpp
@@ -16,10 +16,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "iceoryx_hoofs/testing/mocks/logger_mock.hpp"
-#include "iceoryx_hoofs/testing/test.hpp"
-#include "iox/log/logstream.hpp"
-#include "iox/logging.hpp"
+#include "test_reporting_logstream.hpp"
 
 #include <array>
 #include <bitset>
@@ -28,25 +25,14 @@
 #include <sstream>
 #include <string>
 
+// NOTE: To speed up compilation and the clang-tidy check, the tests for 'iox::string' are split into multiple files
+
 namespace
 {
 using namespace ::testing;
 
-using iox::testing::Logger_Mock;
-
-class LogStreamSut : public iox::log::LogStream
+class IoxLogStream_test : public IoxLogStreamBase_test
 {
-  public:
-    explicit LogStreamSut(iox::log::Logger& logger)
-        : iox::log::LogStream(logger, "file", 42, "function", iox::log::LogLevel::TRACE)
-    {
-    }
-};
-
-class IoxLogStream_test : public Test
-{
-  public:
-    Logger_Mock loggerMock;
 };
 
 TEST_F(IoxLogStream_test, CTorDelegatesParameterToLogger)
@@ -230,434 +216,4 @@ TEST_F(IoxLogStream_test, StreamOperatorLogRawBufferWithNullpointer)
 
     EXPECT_THAT(loggerMock.logs[0].message, StrEq(EXPECTED_DATA));
 }
-
-template <class T>
-class IoxLogStreamHexOctBinIntegral_test : public IoxLogStream_test
-{
-  public:
-    T LogValueLow = std::numeric_limits<T>::lowest();
-    T LogValueMin = std::numeric_limits<T>::min();
-    T LogValueMax = std::numeric_limits<T>::max();
-};
-
-using LogHexOctBinIntegralTypes = Types<uint8_t, int8_t, uint16_t, int16_t, uint32_t, int32_t, uint64_t, int64_t>;
-
-TYPED_TEST_SUITE(IoxLogStreamHexOctBinIntegral_test, LogHexOctBinIntegralTypes, );
-
-template <typename LogType>
-void testStreamOperatorLogHex(Logger_Mock& loggerMock, LogType logValue)
-{
-    LogStreamSut(loggerMock) << iox::log::hex(logValue);
-
-    // we need to check negative numbers in two's complement, therefore make the output value unsigned
-    using TestType = typename std::make_unsigned<LogType>::type;
-    auto outputValue = static_cast<TestType>(logValue);
-
-    std::stringstream ss;
-    ss << "0x" << std::hex << +outputValue; // the '+' is to prevent to interpret the (u)int8_t as char
-
-    ASSERT_THAT(loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(loggerMock.logs[0].message, StrEq(ss.str()));
-}
-
-TYPED_TEST(IoxLogStreamHexOctBinIntegral_test, StreamOperatorLogHex_ValueLow)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "bd47c99a-0808-4a19-bafb-580c65009e0d");
-    testStreamOperatorLogHex(this->loggerMock, this->LogValueLow);
-}
-
-TYPED_TEST(IoxLogStreamHexOctBinIntegral_test, StreamOperatorLogHex_ValueMin)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "ee806b08-0e2a-49fd-b16b-5ad1c8da3150");
-    testStreamOperatorLogHex(this->loggerMock, this->LogValueMin);
-}
-
-TYPED_TEST(IoxLogStreamHexOctBinIntegral_test, StreamOperatorLogHex_ValueMax)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "acfa2bbf-c2e1-42bf-88c6-2888d7d3a42a");
-    testStreamOperatorLogHex(this->loggerMock, this->LogValueMax);
-}
-
-template <typename LogType>
-void testStreamOperatorLogOct(Logger_Mock& loggerMock, LogType logValue)
-{
-    LogStreamSut(loggerMock) << iox::log::oct(logValue);
-
-    // we need to check negative numbers in two's complement, therefore make the output value unsigned
-    using TestType = typename std::make_unsigned<LogType>::type;
-    auto outputValue = static_cast<TestType>(logValue);
-
-    std::stringstream ss;
-    ss << "0o" << std::oct << +outputValue; // the '+' is to prevent to interpret the (u)int8_t as char
-
-    ASSERT_THAT(loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(loggerMock.logs[0].message, StrEq(ss.str()));
-}
-
-TYPED_TEST(IoxLogStreamHexOctBinIntegral_test, StreamOperatorLogOct_ValueLow)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "0163ef2b-7e62-4ded-b093-d77cc4dc360e");
-    testStreamOperatorLogOct(this->loggerMock, this->LogValueLow);
-}
-
-TYPED_TEST(IoxLogStreamHexOctBinIntegral_test, StreamOperatorLogOct_ValueMin)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "68583225-9b47-486d-90da-31f6c1ca7480");
-    testStreamOperatorLogOct(this->loggerMock, this->LogValueMin);
-}
-
-TYPED_TEST(IoxLogStreamHexOctBinIntegral_test, StreamOperatorLogOct_ValueMax)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "c4d104c3-a12a-46e1-8755-cc4b9e0e480c");
-    testStreamOperatorLogOct(this->loggerMock, this->LogValueMax);
-}
-
-template <typename LogType>
-void testStreamOperatorLogBin(Logger_Mock& loggerMock, LogType logValue)
-{
-    LogStreamSut(loggerMock) << iox::log::bin(logValue);
-
-    // we need to check negative numbers in two's complement, therefore make the output value unsigned
-    using TestType = typename std::make_unsigned<LogType>::type;
-    auto outputValue = static_cast<TestType>(logValue);
-
-    auto expectedValue = "0b" + std::bitset<std::numeric_limits<TestType>::digits>(outputValue).to_string();
-
-    ASSERT_THAT(loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(loggerMock.logs[0].message, StrEq(expectedValue));
-}
-
-TYPED_TEST(IoxLogStreamHexOctBinIntegral_test, StreamOperatorLogBin_ValueLow)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "e4e684b7-5bcf-4e8d-8cb1-b6df95c3b37c");
-    testStreamOperatorLogBin(this->loggerMock, this->LogValueLow);
-}
-
-TYPED_TEST(IoxLogStreamHexOctBinIntegral_test, StreamOperatorLogBin_ValueMin)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "f13b3e6a-8f7c-48c2-ae43-35e0a195556e");
-    testStreamOperatorLogBin(this->loggerMock, this->LogValueMin);
-}
-
-TYPED_TEST(IoxLogStreamHexOctBinIntegral_test, StreamOperatorLogBin_ValueMax)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "b583014a-700f-46e3-8b7c-0a128c59598a");
-    testStreamOperatorLogBin(this->loggerMock, this->LogValueMax);
-}
-
-template <class T>
-class IoxLogStreamHexFloatingPoint_test : public IoxLogStream_test
-{
-  public:
-    T LogValueLow = std::numeric_limits<T>::lowest();
-    T LogValueMin = std::numeric_limits<T>::min();
-    T LogValueMax = std::numeric_limits<T>::max();
-};
-
-using LogHexFloatingPointTypes = Types<float, double, long double>;
-
-TYPED_TEST_SUITE(IoxLogStreamHexFloatingPoint_test, LogHexFloatingPointTypes, );
-
-template <typename T>
-constexpr const char* floatingPointFormatSpecifier();
-template <>
-constexpr const char* floatingPointFormatSpecifier<float>()
-{
-    return "%a";
-}
-template <>
-constexpr const char* floatingPointFormatSpecifier<double>()
-{
-    return "%la";
-}
-template <>
-constexpr const char* floatingPointFormatSpecifier<long double>()
-{
-    return "%La";
-}
-
-template <typename LogType>
-void testStreamOperatorLogHexFloatingPoint(Logger_Mock& loggerMock, LogType logValue)
-{
-    LogStreamSut(loggerMock) << iox::log::hex(logValue);
-
-    constexpr uint64_t BUFFER_SIZE{1000};
-    std::array<char, BUFFER_SIZE> buffer{0};
-    // NOLINTNEXTLINE(cppcoreguidelines-pro-type-vararg,hicpp-vararg) required to create a hex formatted float
-    EXPECT_THAT(snprintf(buffer.data(), BUFFER_SIZE - 1, floatingPointFormatSpecifier<LogType>(), logValue), Gt(0));
-
-    ASSERT_THAT(loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(loggerMock.logs[0].message, StrEq(buffer.data()));
-}
-
-TYPED_TEST(IoxLogStreamHexFloatingPoint_test, StreamOperatorLogHex_ValueLow)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "abcad269-7e1b-478a-9535-c9fff9ba1d05");
-    testStreamOperatorLogHexFloatingPoint(this->loggerMock, this->LogValueLow);
-}
-
-TYPED_TEST(IoxLogStreamHexFloatingPoint_test, StreamOperatorLogHex_ValueMin)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "78d3a9c5-3f3e-410a-baba-942c5a1db46e");
-    testStreamOperatorLogHexFloatingPoint(this->loggerMock, this->LogValueMin);
-}
-
-TYPED_TEST(IoxLogStreamHexFloatingPoint_test, StreamOperatorLogHex_ValueMax)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "e87982ae-58e7-4563-99bc-d08484fb60c9");
-    testStreamOperatorLogHexFloatingPoint(this->loggerMock, this->LogValueMax);
-}
-
-template <typename T>
-class Wrapper
-{
-  public:
-    constexpr explicit Wrapper(T value) noexcept
-        : m_value(value)
-    {
-    }
-
-    // NOLINTJUSTIFICATION Implicit conversion is required for the test
-    // NOLINTNEXTLINE(hicpp-explicit-conversions)
-    constexpr operator T() const noexcept
-    {
-        return m_value;
-    }
-
-  private:
-    T m_value{};
-};
-
-using ArithmeticTypes =
-    Types<bool, int8_t, int16_t, int32_t, int64_t, uint8_t, uint16_t, uint32_t, uint64_t, size_t, float, double>;
-
-TYPED_TEST_SUITE(IoxLogStreamArithmetic_test, ArithmeticTypes, );
-
-template <typename Arithmetic>
-class IoxLogStreamArithmetic_test : public IoxLogStream_test
-{
-  public:
-    using type_t = Arithmetic;
-
-    Arithmetic LogValueLow = std::numeric_limits<Arithmetic>::lowest();
-    Arithmetic LogValueMin = std::numeric_limits<Arithmetic>::min();
-    Arithmetic LogValueMax = std::numeric_limits<Arithmetic>::max();
-
-    const Arithmetic ConstLogValueLow = std::numeric_limits<Arithmetic>::lowest();
-    const Arithmetic ConstLogValueMin = std::numeric_limits<Arithmetic>::min();
-    const Arithmetic ConstLogValueMax = std::numeric_limits<Arithmetic>::max();
-
-    static constexpr Arithmetic CONSTEXPR_LOG_VALUE_LOW = std::numeric_limits<Arithmetic>::lowest();
-    static constexpr Arithmetic CONSTEXPR_LOG_VALUE_MIN = std::numeric_limits<Arithmetic>::min();
-    static constexpr Arithmetic CONSTEXPR_LOG_VALUE_MAX = std::numeric_limits<Arithmetic>::max();
-};
-
-template <typename Arithmetic>
-constexpr Arithmetic IoxLogStreamArithmetic_test<Arithmetic>::CONSTEXPR_LOG_VALUE_LOW;
-
-template <typename Arithmetic>
-constexpr Arithmetic IoxLogStreamArithmetic_test<Arithmetic>::CONSTEXPR_LOG_VALUE_MIN;
-
-template <typename Arithmetic>
-constexpr Arithmetic IoxLogStreamArithmetic_test<Arithmetic>::CONSTEXPR_LOG_VALUE_MAX;
-
-template <typename T>
-std::string convertToString(const T val)
-{
-    return std::to_string(val);
-}
-
-template <>
-std::string convertToString<float>(const float val)
-{
-    std::stringstream ss;
-    ss << val;
-    return ss.str();
-}
-
-template <>
-std::string convertToString<double>(const double val)
-{
-    std::stringstream ss;
-    ss << val;
-    return ss.str();
-}
-
-template <>
-std::string convertToString<bool>(const bool val)
-{
-    return std::string(val ? "true" : "false");
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_ValueLow)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "31f1504a-9353-4c46-9c8b-d7e430b07bd6");
-    LogStreamSut(this->loggerMock) << this->LogValueLow;
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->LogValueLow)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_ValueMin)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "e784ceb9-1e23-4e95-b667-855835897717");
-    LogStreamSut(this->loggerMock) << this->LogValueMin;
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->LogValueMin)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_ValueMax)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "3bff0182-07ad-4c7a-b8b7-3950a8aa9f4e");
-    LogStreamSut(this->loggerMock) << this->LogValueMax;
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->LogValueMax)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_WrappedAndImplicitlyConverted_ValueLow)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "86dccc28-0007-4ab3-88c6-216eeb861fef");
-    using type_t = typename TestFixture::type_t;
-    LogStreamSut(this->loggerMock) << Wrapper<type_t>(this->LogValueLow);
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->LogValueLow)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_WrappedAndImplicitlyConverted_ValueMin)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "72bc51fe-a23f-4c7d-a17f-0e4ad29bee9f");
-    using type_t = typename TestFixture::type_t;
-    LogStreamSut(this->loggerMock) << Wrapper<type_t>(this->LogValueMin);
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->LogValueMin)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_WrappedAndImplicitlyConverted_ValueMax)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "49e79c8b-98c0-4074-aed1-25293bf49bd0");
-    using type_t = typename TestFixture::type_t;
-    LogStreamSut(this->loggerMock) << Wrapper<type_t>(this->LogValueMax);
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->LogValueMax)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_ConstValueLow)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "65cfbc9b-a535-47fa-a543-0c31ba63d4ba");
-    LogStreamSut(this->loggerMock) << this->ConstLogValueLow;
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->ConstLogValueLow)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_ConstValueMin)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "fba70497-e252-4458-b00e-2dad8b94b8c8");
-    LogStreamSut(this->loggerMock) << this->ConstLogValueMin;
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->ConstLogValueMin)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_ConstValueMax)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "e5e28a6e-4321-4030-b53e-90089b3ee9b9");
-    LogStreamSut(this->loggerMock) << this->ConstLogValueMax;
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->ConstLogValueMax)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_WrappedAndImplicitlyConverted_ConstValueLow)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "77cb68ac-bffb-4cf5-a4c2-119e5aaae307");
-    using type_t = typename TestFixture::type_t;
-    LogStreamSut(this->loggerMock) << Wrapper<type_t>(this->ConstLogValueLow);
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->ConstLogValueLow)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_WrappedAndImplicitlyConverted_ConstValueMin)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "3b042ec2-0a6d-4d24-9a12-dab81736847f");
-    using type_t = typename TestFixture::type_t;
-    LogStreamSut(this->loggerMock) << Wrapper<type_t>(this->ConstLogValueMin);
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->ConstLogValueMin)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_WrappedAndImplicitlyConverted_ConstValueMax)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "83194cc1-3c2a-4c0e-a413-972a95f36b66");
-    using type_t = typename TestFixture::type_t;
-    LogStreamSut(this->loggerMock) << Wrapper<type_t>(this->ConstLogValueMax);
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->ConstLogValueMax)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_ConstexprValueLow)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "e9688979-d209-4718-9810-49684fdd9261");
-    LogStreamSut(this->loggerMock) << this->CONSTEXPR_LOG_VALUE_LOW;
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->CONSTEXPR_LOG_VALUE_LOW)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_ConstexprValueMin)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "f6799599-582a-454c-85b8-b2059a5d50c6");
-    LogStreamSut(this->loggerMock) << this->CONSTEXPR_LOG_VALUE_MIN;
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->CONSTEXPR_LOG_VALUE_MIN)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_ConstexprValueMax)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "4a6dc777-a53b-4a42-9ab1-e1da893ad884");
-    LogStreamSut(this->loggerMock) << this->CONSTEXPR_LOG_VALUE_MAX;
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->CONSTEXPR_LOG_VALUE_MAX)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_WrappedAndImplicitlyConverted_ConstexprValueLow)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "3025b6e3-3abd-49e8-8d3c-f6fa0c164011");
-    using type_t = typename TestFixture::type_t;
-    LogStreamSut(this->loggerMock) << Wrapper<type_t>(this->CONSTEXPR_LOG_VALUE_LOW);
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->CONSTEXPR_LOG_VALUE_LOW)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_WrappedAndImplicitlyConverted_ConstexprValueMin)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "c49ad58d-948a-499a-9e2b-8bbdeb9b5c5e");
-    using type_t = typename TestFixture::type_t;
-    LogStreamSut(this->loggerMock) << Wrapper<type_t>(this->CONSTEXPR_LOG_VALUE_MIN);
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->CONSTEXPR_LOG_VALUE_MIN)));
-}
-
-TYPED_TEST(IoxLogStreamArithmetic_test, StreamOperator_WrappedAndImplicitlyConverted_ConstexprValueMax)
-{
-    ::testing::Test::RecordProperty("TEST_ID", "8c7637ce-8155-4edc-88e2-a155169934f2");
-    using type_t = typename TestFixture::type_t;
-    LogStreamSut(this->loggerMock) << Wrapper<type_t>(this->CONSTEXPR_LOG_VALUE_MAX);
-
-    ASSERT_THAT(this->loggerMock.logs.size(), Eq(1U));
-    EXPECT_THAT(this->loggerMock.logs[0].message, StrEq(convertToString(this->CONSTEXPR_LOG_VALUE_MAX)));
-}
-
 } // namespace
